Refactor Azure with Windows
