+++ tensorflow_io/core/kernels/azfs/azfs.cc
@@ -21,11 +21,11 @@ limitations under the License.
 #include "tensorflow/core/lib/io/path.h"
 #include "tensorflow/core/lib/strings/str_util.h"
 #include "tensorflow/core/platform/env.h"
-#include "tensorflow_io/core/azure/azfs/azfs.h"
-#include "tensorflow_io/core/azure/azfs/azfs_client.h"
-#include "tensorflow_io/core/azure/azfs/azfs_random_access_file.h"
-#include "tensorflow_io/core/azure/azfs/azfs_readonly_memory_region.h"
-#include "tensorflow_io/core/azure/azfs/azfs_writable_file.h"
+#include "tensorflow_io/core/kernels/azfs/azfs.h"
+#include "tensorflow_io/core/kernels/azfs/azfs_client.h"
+#include "tensorflow_io/core/kernels/azfs/azfs_random_access_file.h"
+#include "tensorflow_io/core/kernels/azfs/azfs_readonly_memory_region.h"
+#include "tensorflow_io/core/kernels/azfs/azfs_writable_file.h"
 
 #include "blob/blob_client.h"
 #include "storage_account.h"
