+++ tensorflow_io/core/kernels/azfs/azfs_client.cc
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow_io/core/azure/azfs/azfs_client.h"
+#include "tensorflow_io/core/kernels/azfs/azfs_client.h"
 #include "logging.h"
 
 namespace tensorflow {
