+++ tensorflow_io/core/kernels/azfs/azfs_test.cc
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "tensorflow_io/core/azure/azfs/azfs.h"
+#include "tensorflow_io/core/kernels/azfs/azfs.h"
 #include "gtest/gtest.h"
 
 #define EXPECT_OK(val) EXPECT_EQ(val, Status::OK())
